requirements.txt
# BrasileirãoGPT - Dependências do Projeto

# Framework web
streamlit>=1.31.0

# LangChain e dependências
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# OpenAI
openai>=1.12.0

# Utilitários
python-dotenv>=1.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Opcional: para usar outras funcionalidades
# serpapi  # Para busca na web (requer API key)
# wikipedia  # Para buscas na Wikipedia
